Move access control backends into courseware
Because it depends on AccessResponse class
