+++ lms/djangoapps/courseware/access.py
@@ -37,7 +37,7 @@ from courseware.access_utils import (
 from courseware.masquerade import get_masquerade_role, is_masquerading_as_student
 from lms.djangoapps.ccx.custom_exception import CCXLocatorValidationException
 from lms.djangoapps.ccx.models import CustomCourseForEdX
-from lms.lib.access_control_backends import access_control_backends
+from lms.djangoapps.courseware.access_control_backends import access_control_backends
 from mobile_api.models import IgnoreMobileAvailableFlagConfig
 from openedx.core.djangoapps.content.course_overviews.models import CourseOverview
 from openedx.core.djangoapps.external_auth.models import ExternalAuthMap
