+++ tox.ini
@@ -121,11 +121,10 @@ commands =
         lms/djangoapps/course_api/ \
         lms/djangoapps/course_blocks/transformers/tests/test_load_override_data.py \
         lms/djangoapps/courseware/tests/test_access.py \
-        lms/djangoapps/courseware/tests/test_access_control_backends_integration.py \
+        lms/djangoapps/courseware/tests/test_access_control_backends.py \
         lms/djangoapps/grades/tests/integration/test_events.py \
         lms/djangoapps/instructor/tests/test_certificates.py::CertificatesInstructorApiTest \
         lms/djangoapps/verify_student/tests/test_services.py  \
-        lms/lib/tests/test_access_control_backends.py \
         openedx/core/djangoapps/appsembler \
         openedx/core/djangoapps/site_configuration/tests/test_tahoe_changes.py \
         openedx/core/djangoapps/user_api/accounts/tests/test_utils.py::CompletionUtilsTestCase
